rpk: add cluster view version to maintenance status and broker list
