+++ src/go/rpk/pkg/api/admin/api_broker.go
@@ -12,11 +12,9 @@ package admin
 import (
 	"fmt"
 	"net/http"
-	"sort"
 )
 
 const brokersEndpoint = "/v1/brokers"
-const clusterViewEndpoint = "v1/cluster_view"
 
 type MaintenanceStatus struct {
 	Draining     bool `json:"draining"`
@@ -28,11 +26,6 @@ type MaintenanceStatus struct {
 	Failed       int  `json:"failed"`
 }
 
-type ClusterView struct {
-	Version int      `json:"version"`
-	Brokers []Broker `json:"brokers"`
-}
-
 // Broker is the information returned from the Redpanda admin broker endpoints.
 type Broker struct {
 	NodeID           int                `json:"node_id"`
@@ -41,16 +34,6 @@ type Broker struct {
 	IsAlive          *bool              `json:"is_alive"`
 	Version          string             `json:"version"`
 	Maintenance      *MaintenanceStatus `json:"maintenance_status"`
-}
-
-// Brokers queries one of the client's hosts and returns the list of brokers.
-func (a *AdminAPI) Brokers() ([]Broker, error) {
-	var cv ClusterView
-	defer func() {
-		sort.Slice(cv.Brokers, func(i, j int) bool { return cv.Brokers[i].NodeID < cv.Brokers[j].NodeID }) //nolint:revive // return inside this deferred function is for the sort's less function
-	}()
-	err := a.sendAny(http.MethodGet, clusterViewEndpoint, nil, &cv)
-	return cv.Brokers, err
 }
 
 // Broker queries one of the client's hosts and returns broker information.
